Move the specs to the separate folder
